requirements.txt
appdirs==1.4.3
Django==1.9.7
django-ckeditor==5.2.1
django-cuser==2017.3.16
django-datetime-widget==0.9.3
django-multiupload==0.5.2
django-nocaptcha-recaptcha==0.0.19
django-tinymce==2.6.0
django-widget-tweaks==1.4.1
easy-thumbnails==2.4.1
flynsarmy-paginator==0.0.1
modeltranslation==0.25
olefile==0.44
packaging==16.8
Pillow==4.1.0
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
